move clear predictions button to the bottom of category list, minor style changes, resolve #254
