+++ src/components/CategoriesList/CategoriesList.tsx
@@ -1,5 +1,5 @@
 import { CollapsibleList } from "../CollapsibleList";
-import React from "react";
+import React, { useEffect } from "react";
 import { useSelector } from "react-redux";
 import { CategoryListItem } from "../CategoryListItem";
 import { Category } from "../../types/Category";
@@ -8,10 +8,20 @@ import {
   createdCategoriesSelector,
   unknownCategorySelector,
 } from "../../store/selectors";
+import { predictedSelector } from "store/selectors/predictedSelector";
+import { PredictionVisibility } from "components/PredictionsVisibility/PredictionVisibility";
 
 export const CategoriesList = () => {
   const categories = useSelector(createdCategoriesSelector);
   const unknownCategory = useSelector(unknownCategorySelector);
+
+  const predicted = useSelector(predictedSelector);
+  const [showPredictionVisibilityMenu, setShowPredictionVisibilityMenu] =
+    React.useState<boolean>(true);
+
+  useEffect(() => {
+    setShowPredictionVisibilityMenu(predicted);
+  }, [predicted]);
 
   return (
     <CollapsibleList primary="Categories">
@@ -29,6 +39,7 @@ export const CategoriesList = () => {
           />
         );
       })}
+      {showPredictionVisibilityMenu && <PredictionVisibility />}
       <CreateCategoryListItem />
     </CollapsibleList>
   );
